+++ packages/client-sdk/src/index.ts
@@ -13,4 +13,7 @@ export { Events as IFrameEvents } from "./iframe/blocks.js";
 
 export { MashEvent } from "./events.js";
 
-export { WalletButtonFloatSide } from "./api/routes.js";
+export {
+  WalletButtonFloatSide,
+  WalletButtonFloatPlacement,
+} from "./api/routes.js";